requirements.txt
altgraph==0.17.4
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyperclip==1.9.0
pytesseract==0.3.13
pywin32-ctypes==0.2.3
setuptools==75.1.0
